# qemu: add patch improving 9p performance