@K900 2022-10-13
Before:
./result/bin/nixos-test-driver  1660,85s user 156,92s system 115% cpu 26:14,37 total

After:
./result/bin/nixos-test-driver  201,56s user 142,66s system 262% cpu 2:11,06 total

Just a casual 13x speedup.