@kvtb 2022-12-17
Wow.
So, GitLab test could be enabled again.
It timed out because of too many small Ruby files